With the help of another Array, fitting currencies, according the entered String, can be selected and will be shown in the menu OR if only one currency is found, it will be added directly to the menu. Also added variables for the menu-dynamic to allow changes when both currencies are selected. Started to integrate the Calculator class.

+++ src/main/java/de/hdm_stuttgart/mi/sd1/Calculator.java
@@ -2,7 +2,7 @@ package de.hdm_stuttgart.mi.sd1;
 
 public class Calculator {
 
-    float result;
+    double result;
 
     /**
      * Algorithm to convert an amount of one currency to another with the help of their SDR values
@@ -13,7 +13,7 @@ public class Calculator {
      * @return The converted amount
      */
 
-    public float convertingAmount (float amount, float buySDR, float sellSDR) {
+    public double convertingAmount (double amount, double buySDR, double sellSDR) {
 
         result = (amount/buySDR)*sellSDR;
         return result;
